dash: refresh the MPD less often for SegmentTimeline-based indexes
Previously, we updated the MPD as soon as a wanted position is after the current manifest. Now, we check that the end of what is wanted is at least at end of the index + duration of a segment (based on the last segment). This might not be the best logic (refresh can still happen a lot) but its much better than before.