dash: give the clockoffset to the parser instead of just telling it we have one
The previous behavior led to a bug for dynamic SegmentTemplate-based contents without a single SegmentTimeline. This is linked to the fact that we only fetch the UTCTiming the first time the Manifest is parsed and pass for the other ones (as we already have it). The problem is that in a certain condition - the one aforementioned - we use that timing to calculate the maximum possible position (basically the live time here). Without it, we used Date.now() instead (which may be completely off).

The new Behavior now is to actually communicate the clock offset to the parser if we have one. The MPD parser may then use it directly