buffers: when choosing the needed segments take into consideration more of current segments that actually needed
This is a precaution to avoid the situation where an anounced segment is slightly shorter than initially anounced. When paused just after the real segment, we may not include it in our list of current segments as neededRange.start would be higher than its real inferred end. In that case, we would download the same segment in a loop