custom_source_buffers/text: fix text track garbage collection and re-insertion afterwards.
The following changes have been brought in this fix:

  1. we are much less aggressive with text track garbage collection: When
removing in the middle of a cue in a given segment, we update the start
of the segment but keep the corresponding cue (it was removed before).
It will just last less time than initially set for (but that's what we
want, based on the removal)

  2. Some fixes have been brought in the remove function of the HTML text track BufferManager.
     The remove function missed some possible cases.

  3. Be exhaustive for the placement possibilities in the `insert`
function of the HTML text track BufferManager. Using GC means that sub-parts of
a segment can appear in a buffer. This can take almost any form when
comparing a new segment which should be put in its place. The `insert`
function has been refactored and should (normally) handle all possible
cases now.