utils/eme: stop relying on hashBuffer directly to better handle hash collision
Fix hash collision management in the storage of EME initialization data. The previous behavior could theorically lead to weird behavior such as some initialization data to be ignored or the wrong persistent session to be loaded.\n The current implementation rely on a very simple array of tuples which should work and is simple enough but could be problematic as the current list of session grows.