eme: completely rollback to always use an array and linear search to keep a single MediaKeySession per initialization data
After first implementing a sort of HashMap, which seems more adapted to the current situation, I found out that the order of insertion had an importance. My implementation did not consider that, and adding ordering into it would have made it much more complex.\n As we're talking about very few concurrent MediaKeySessions here, I decided that implementing the simpler and less performant solution of using a simple JS Array with an added initData hash for faster comparison was a good enough compromise.