eme: add getAndReuse method to the LoadedSessionsStore to allow LRU cache eviction
The previous cache replacement policy was to replace the MediaKeySession initially created at the earliest time (in a FIFO manner). It would actually make more sense here to replace the least recently used instead. For example, if a MediaKeySession created long ago is re-used before reaching the cache limit, we don't want to remove its entry from the LoadedSessionsStore before those linked to contents it played just before it - just because its creation time was before.

More simply put, we could say that our replacement policy is now based on the oldest usage time, whereas it was previously based on the oldest creation time.