eme: in the InitDataStore, perform lookup starting from the end of the Array
When wanting to look for a value stored thanks to the InitDataStore, we previously went into every stored values from first to last.
However for most initData-related usecases, it is generally more efficient to begin from the last entry to the first one.

For example, when zapping between two live encrypted channels, you generally will only encounter the last two initData consecutively. Re-browsing through the first stored data would just be wasteful here.
This might however not be better for all usecases, but it is still a saner default.