dash/parsers: use AdaptationSet@selectionPriority to prioritize AdaptationSets
This is for now simply done by sorting AdaptationSets from a given Period just before they are actually parsed in the `parseAdaptationSets` function. I chose to do it before as it is much simpler that way but it might be less error-prone to do it at the end, as we wouldn't need to consider the array order until then.

The main video adaptation is still considered to be the first for now and has priority over selectionPriority. I don't know what we should do in that case