init: refresh the Manifest even if we dont have a Manifest URL
Under the previous code, we aborted any Manifest refresh operation if we did not find an URL associated to the Manifest. This might seem logical at first, but we may also have a custom `manifestLoader` defined through the corresponding loadVideo option, which make the presence of the URL completely optional.

The real use cases are for when the Manifest is available locally, most of all under the `local` transport to be able to play content offline. In such contents, we never have an URL, only a custo `manifestLoader` which will retrieve the stored contents directly from memory or whatever browser storage API is available (IndexedDB, custom APIs...).