utils: make much more use of the assertUnreachable function, to check branch exhaustiveness
I added multiple `assertUnreachable` checks to branch-heavy places in
our code.

Basically, this is mostly to ensure at compile-time that every possible
cases have been handled. Here, I'm most-of-all thinking about our
event-type discriminated unions which rely on big switch statement where
different events correspond to different actions taken in response.

Not doing this assertion could lead to bugs. For example, we could
decide to add an event but forget to update a switch elsewhere, which
might receive that event and perform a default action on it we do not
want it to do.

I actually found an example of that in our code. There is a minor bug
where the new (v3.20.1) `cleaning-old-session"` and
`"cleaned-old-session"` events were not properly ignored in
`eme_manager.ts`.
Here we could be binding again a `SessionEventsListener` to a
MediaKeySession which has just been removed.

It could be avoided by adding a call to `assertUnreachable` in the
`default` branch of the corresponding `switch`. When we add new events
that are concerned by that `switch`, TypeScript alerts us that this call
is not legal, redirecting our attention towards that new problem created
due to a modification which was first thought as irrelevant to that part
of the code.

Still, this `assertUnreachable` trick is not ideal, as it creates the
same function in the generated JS code. This means that we do not get
all optimizations we could have had, had the check not been added. This
is better explained in the TypeScript issue I have opened here:
https://github.com/microsoft/TypeScript/issues/38881