source-buffers: Do not create a video SourceBuffer on an audio element
The RxPlayer can be bound to any HTMLMediaElement (for now this just
means audio and video elements). Previously the RxPlayer downloaded all
types of buffers ("video", "audio" and depending on the build "text" and
"image") regardless of what element was used.

However it's mainly a waste of network resources to download video on an
audio element, as those cannot render the video we will push to its
SourceBuffer.

This PR does a special treatment when encountering an HTMLAudioElement.
Only on this case, no video will be downloaded.

I did that by updating the place where `getBufferTypes` - which returns
the complete list of supported buffer types - is defined.
Previously at top-level, it is now a method of the SourceBufferStore
(which makes sense as it already contained similar methods such as
`isNative`, and as it has to know the list of SourceBuffer to create).
I also had to update its `waitForUsableSourceBuffers` as we don't want
to wait for a video SourceBuffer to be created in this case.

I chose to checking for an "audio" element to not include the "video"
buffer type and not the other way around (checking for a "video" element
to include it) as it gives a nice behavior in the case where an
HTMLMediaElement is neither (for now, I don't think this is possible on
any platform but it's a possibility).
On every other possible HTMLMediaElement, the previous behavior (both
audio and video) will be active.

This feature was inspired from what other player does (most
notably the Shaka-Player).