demo: don't depend on babel/polyfill for the demo anymore as it is deprecated
As seen in another PR (#729) the @babel/polyfill was only a shortcut to both core-js and the regenerator-runtime, which are also two different things.

Understandably, @babel/polyfill is now deprecated in profit of importing both directly.

Despite the appearance I add no dependencies to the RxPlayer here, as what I added was already a dependency of @babel/polyfill.
Consequently, I even removed one (which is... @babel/polyfill!).

This seems to be lead toward a less readable code when importing both in the demo but it actually kind of make much more sense when the time comes where you have to understand what all this does.

I would still have preferred not have to do with a regenerator all together (still don't quite know why it's not added directly by babel during transpilation).