eme: move responsibility of how many MediaKeySession to clean up from stores to the the caller
Previously, it was directly the clean-up code which used the value from
the configuration.

The problem was that we usually want to make place for the next
MediaKeySession when calling those functions.
So i made them remove one more MediaKeySession.

However, this led to a less understandable code. Why would those
functions, which apparently just perform "clean-up" remove one more
MediaKeySession that they should?

Even more ridiculous, the loop clearing up sessions in that functions
actually didn't remove that supplementary MediaKeySession. It was thus
both un-readable and buggy (well, we would have one more MediaKeySession
than the limit at the end).

So I decided to move the responsibility to declare the wanted maximum
MediaKeySession to the caller. The code of the function itself just
remove entries to stay in that limit. This makes much more sense to me.

Also, I saw another thing that looked like a bug (but was not), in that
we were re-computing the loop condition (from the number of entries and
its maximum limit) at each iteration, despite the fact that the loop
itself seemed to remove entries.
However we were safe in that removing those entries was a lazy iteration
only done at subscription (which is done after this function has
returned).

Last but not least, the condition in `cleanOldStoredPersistentInfo` was
also backward.

I'm feeling that we really need to had some testing in there.
Integration tests are limited to non-encrypted contents for now but we
could still work with people at the UI-layer to make some with Canal+
contents.

In this case here, we could also just write simple unit tests. I'm going
to do that now.