demo: now use "!" as a key separator in generated links
Fixes #758.

Before we used a backslash ("\") character as a key separator in
generated link but I found out those needed to be percent-encoded in a
valid URL.

We had no problem when directly linking to that URL but I found out,
after sharing that link on outlook, that sharing that link there
"broke" the format of that string as outlook percent-encode that URL
when it is shared through e-mails (this seems to be a side effect of
their "safelink" protection scheme).

When getting links that where percent-encoded, the demo page didn't know
what to do and thus acted as if no content was set through the URL (as
a resilience measure).

By using now "!", a perfectly valid character in the fragment of an URL,
we can now avoid that issue.

Note that the demo page is still compatible with URL with "\"
characters, so links created with older versions still work.