fetchers: add the concept of "high priority" segment requests
Solves #752.

In our segment request scheduling logic, we have a notion of priority.
Each new segment request has an associated priority number. The lowest
is that number (starting at 0), the highest is that request's priority
over other segment requests.

When multiple segment requests are needed (generally we are talking
about segment of different types here - like audio, video... - which are
downloaded concurrently), the "segment fetchers" (the part of the
RxPlayer scheduling and asking for segment requests) will first start
the request with the highest priority. Once done, it will start the
request with the next highest priority and so on.

Because the player is playing at the same time that the requests are
performed, it is possible that the priority of different segment
requests evolve (for example, because a segment that was before far
away is now very close).
In that case the segment request's priority is updated accordingly.

However, until now, a segment request priority only affected when that
request will be "started". Once started, a priority update on that
request had no effect.

This has been the behavior of the RxPlayer for some time until this PR.
I changed a little that logic because in some rare cases (like the one
the #752 issue is about), segment requests of very high priority - like
segments for the current position, which are necessary before playing -
could be needed at the same time that segment requests of very low
priority are pending (already started).

In that last situation, it would be sad to limit the speed at which we
fetch those high priority segments, because low priority ones are using
too much network bandwidth.

For that last situation, I choose to add the concept of both a "high
priority" step and a "low priority" step.

When segment requests with "high priority" are scheduled, all "low
priority" requests are interrupted. They are put back in the waiting
queue, to wait for their turn to be started.

But this brings a strange behavior in the point of view of the
`Buffers`: to restart from scratch a request that was already started
and might had already communicated chunks.
For that issue, I make it explicit when an interrupt happens by adding
another event, `"interrupted"`, which will be sent any time one of this
situation happens.

There was a last problem - and not a small one - which is that the
current logic has a flaw which is akin to a race condition (though it is
not really one).

The problem is that in our current logic, when a `RepresentationBuffer`
ask for multiple segment requests, it usually goes the following way:
  1. Ask the segment fetchers for the most needed segment (with its
     associated priority).
  2. Once it has been fully downloaded, ask the next most needed segment
     (also with is associated priority).
  3. Go back to 2.

The problem is that the request has finished in the segment fetchers
point of view just before (2) is reached. As such, it has already
scheduled the next request. This new request might have a lower
priority than the one scheduled in (2) but the segment fetchers doesn't
know about it yet.

In a perfect situation, the segment fetchers should always compare the
priority of all needed segments. As such the `RepresentationBuffer`
should be able to tell about the next segment before the request for the
current one has ended.

For this issue, I have two ideas, one short term (that I implemented
there) and one longer term (that I'm working on in another branch).

For the short term solution: we can add another event, today called
"ended", which will be sent just before a segment request completes. By
adding the next request synchronously on that event, we can ensure
that the next request is scheduled before the segment fetchers makes its
choice.

This is what I implemented here. This solution has still small issues
where we would still have a similar problems when the Buffers interrupt
the current segment request before it ends, such as when seeking to a
completely different position.

For the long term solution: I would like to make the segment fetchers
aware of all the "wanted segment list", so they can take the task of
scheduling EVERY segments that are needed at a given time, and not
just the next segment of a given type.

This basically means moving logic from the `RepresentationBuffer` to the
segment fetchers. I feel that we can do a much more "smart" logic in the
latter without fearing to complexify the code too much as we are in a
smaller, well-contained, unit of code.