buffers: simplify code readability for the `filterGarbageCollectedSegments` and fix enc check issue
There was an issue in the previous `filterGarbageCollectedSegments` code
where we would consider the end of a segment to be garbage collected
even if a segment can be found exactly after it (its start equal our
first segment's end).

This is not wanted behavior and was the source of an issue in firefox
with the multi-period DASH test stream.

Here, we might think that a given audio segment at the end of the first
period is garbage collected, in which case we would just request it in a
loop.