features: avoid circular dependency in `src/features`
This is a small fix to avoir a circular dependency in the `src/features`
directory.

Basically, the `features` object, which, was declared in
`src/features/index.ts`, was imported by `src/features/add_features.ts`.
This is problematic because `src/features/index.ts` also imports
`src/features/add_features.ts`.

There was no real reason to declare the `features` object in `index.ts`
other than convenience. In this commit, I move it to its own
`features_object.ts` file.

This issue led to no build error that I know of, but it is still an
ugly avoidable situation.