Buffers: set priority of the init segment as the same than the most wanted segment
Originally, the priority of the init segment was always set to 0. This
is not always wanted.

We would be using network resources for segment(s) that may not be
immediately needed.

A much better default behavior in my opinion is to set it the priority
of the most wanted segment for that type instead.