manifest: do not update manifest uris on partial updates
Fixes #775.

A "partial update" (that is an update with a shorter version of the
Manifest to allow parsing optimizations) of a Manifest wrongly updated
the `uris` property of the Manifest.

This is not what we want to do because it means that for the RxPlayer
the new "true" URL of the Manifest is the one taken from the shorter
version.

This has no effect right away, but after some time we might be in a
situation where the RxPlayer want to re-perform a "full update" with
what it thinks is the URL of the "normal" Manifest (which is in reality
still the URL of the shorter version). This usually happens when the
RxPlayer wants to make sure its local version of the Manifest is still
synchronized to the true version from the server.

When doing so, it resets the current content as if we were playing the
shorter version. The most visible difference is that the minimum
position appears now to be too high.

I also re-ordered other properties updates in alphabetical order here!