tests: raise tolerance on memory tests again
After a lot of false negatives, I found that the memory taken for our
memory tests appear to be very highly variable.

I doubled the step at which we fail due to too much memory usage.

Obviously, this is not the commit I am the most proud of.

To justify this update to myself, I told myself that a "TRUE" memory
leak would be much more visible than just some MB not yet GCed

For further investigations, we could also check that waiting a little
more time before looking at the heap's size changes something.