tests: add error message to the init segment integration test
This error is susceptible to fail when we break init segment requests,
like I did when I decided to request the init segment only when media
segments are needed (which breaks SegmentBase DASH contents as those
need the index segment - which for now is part of the init segment in
the RxPlayer - before we know about the media segments).

To make it more visible to see which test has failed, I added a more
legible error message than the simple "expected 0 to equal 2" that we
had before.

I guess it could be considered a best practice to add error message in
integration tests in general to quickly understand which step fails.