manifest/buffers: add `isInitialized` method to a RepresentationIndex
I recently made a code modification (not yet merged in master) that
broke the playback of SegmentBase-based Representations.

Basically, the code modification among other things was only scheduling
an init segment request when at least one segment from the same
Representation is needed.

This made sense because we might not want to waste resources for an init
segment that is not even needed, but it actually breaks when we consider
the notion of the "index segment" which is today part of the init
segment in the RxPlayer.

The "index segment" is a segment that the RxPlayer has to load which
contains the list of media segment for a Representation. Before that
index segment is loaded, we don't have any list of segment and as such,
the `getSegments` method will always return an empty array.

Because `getSegments` returns an empty array, my code modification made
it sure that we don't download the init segment. And because the init
segment contains the index segment, `getSegments` would always return an
empty array. Result: the Representation could not play.

To work-around that explicitely, I added the notion of an
"initialized" RepresentationIndex.

An initialized RepresentationIndex is a RepresentationIndex with a known
list of segments. Most RepresentationIndex implementations are always
initialized, because they have no concept of an index segment.

The only un-initialized RepresentationIndex for now is the SegmentBase
one, which become initialized as soon as its `_addSegments` method adds
at least a single segment to the timeline.

I also updated the logic in the `RepresentationBuffer`.
Now, if an unitialized RepresentationIndex is encountered, it will
request its init segment and nothing else with a priority corresponding
to the start of the corresponding Period.

If no init segment exists in that RepresentationIndex or if one was
already loaded, we log the problem through a `log.warn`.

On that last note, I found it weird that the initialization of a
RepresentationIndex happens implicitely when an init segment is loaded.
This is something that might not seem natural and it adds a notion of
responsibility: who has to initialize the RepresentationIndex.

For now, it is the `transports` but I would be inclined now to think
that it could also be the buffers.

Also, we could add an `initialize` method on the RepresentationIndex -
accepting the index segment - to explicitely have an initialization
process.

I did not perform those last modifications yet as I'm unsure of whether
or not they make sense.