code: remove all `ICompatMediaKeySystemConfiguration` occurences
There was a small typing issue with TypeScript and the
MediaKeySystemConfiguration type.

Basically, the "sessionTypes" attribute was not present. Because we
needed it, we had to create our own version called
`ICompatMediaKeySystemConfiguration` which included it.

We also opened an issue on the TypeScript repo at that time, registered
on github as issue #19189. They proposed that we fix the problem
ourselves which I wanted to do, but I was quickly lost on how to
properly do so and thus I just abandonned.

Time passed and the problem was not resolved. I just added "XXX TODO"
comments on top oe the `ICompatMediaKeySystemConfiguration` to always
remind me to check if the issue was still not resolved !

As I checked again today, the issue seems to be resolved. This commit
remove any occurence of `ICompatMediaKeySystemConfiguration` and replace
it with the TypeScript `MediaKeySystemConfiguration`.

I also replaced all occurence of `ICompatMediaKeySystemAccess` by
`MediaKeySystemAccess` as the only difference was that the former
returned `ICompatMediaKeySystemConfiguration` in its `getConfiguration`
method.