buffers: fix race condition happening in tests and simplify AdaptationBuffer code
Even if it didn't reflect with manual testing, some integration tests
failed due to what looks like a race condition triggered in the RxJs
library.

This appeared to be due to the fact that we rely on the same ABR
estimation Observable at two different places in the AdaptationBuffer:
  - immediately on the same scope for e.g. emitting new bandwidth
    estimation event to the outside or to create the first
    RepresentationBuffer
  - in the inner scope of an inner Observable which is the one where
    `RepresentationBuffer`s are recursively created.

After a lot of different drafts, I found out that using a
BehaviorSubject, which does not depend on the asapScheduler that the ABR
estimation Observable rely on both fixes the problem and looks clean
enough: it is simply the last estimation emitted, and reemits each time
it changed.

I profited from that change to make some other minor code changes, like
moving the representation estimator creation logic to its own file.