parsers/isobmff: Support box length defined on 8 bytes
I recently realized that we always assume in the RxPlayer that a box
length is encoded on the first 4 bytes of an ISOBMFF box.

This is not always true, there is actually 2 special values an ISOBMFF
box size attribute can have:
  - `0` the box goes on until the end of the whole data
  - `1` the box size is actually defined on 8 bytes, after the name

To be perfectly honest, the fact that we didn't handle that is not
dramatic. An ISOBMFF box size which cannot be defined on 4 bytes would
mean that the box is much larger than 4GB - this is much higher than
what we ever seen.

But still, it is theorically possible. Moreover, nothing seems to forbid
the size being set on 8 bytes beside being "storable" on just 4.

Updating the RxPlayer's code so it can handle box length on 8 bytes was
a lot more complex than I initially thought because much of our
ISOBMFF-related code actually directly considered the size to be on 4
bytes and used this `4` value directly inline.

Being able to generalize the new behavior elegantly led me to factorize a
lot more code.
For now, it also added some peformance regressions as some operations are
performed multiple times (to improve clarity) and as some boxes are now
copied instead of mutated in place.
Though I don't think that this will be too visible for the end-user
(we're not talking about performance-sensitive code at all here).

The code updated is a little touchy, I read, tested and re-read the code
but I'm still not that confident. A mistake here will in most case
completely break playback for some contents. This is something to keep
in mind.