dash: Check if bytes are missing from segment when the `checkMediaSegmentIntegrity` option is set
This is a PR done reluctantly to work-around some problems we have with
VoD contents at Canal+ Group.

This PR adds a new behavior for the checkMediaSegmentIntegrity option
(given as a transportOptions to a loadVideo call) for DASH contents: it
now also checks that the segment requested through byte-range requests
did not miss out some bytes due to a wrong Range headers in the
request.

Now the segment loaders can emit a new `"direct-retry"` event, which
indicates that they are immediately re-performing the same segment
request (with only minor differences such as different headers).

This event is send when a requested segment with a linked byte-range is
found to be incomplete due to that byte-range being too short. The
request is re-performed that way with a "bigger" byte-range indicated.