abr: better document ABR filters code and fix bitrate filtering issue
There was a small issue in `filter_by_bitrate.ts` where we could be
filtering out all Representations.

In the scope of that function, it looks fine but when considering how
the outside code considers what that function does, it's less clear.

I fixed an issue some time before without ever understanding when the
original bad conditions happened. I now thinks that it has its source
from that situation.

I updated `filter_by_bitrate` code and documentation so it is now
clearer that if no Representation has a lower or equal bitrate, we still
consider those with the lowest bitrates.