abr: better document the `RepresentationEstimator` function
While working on segment requests and its related events (network
metrics, events about request progressions etc.) I noticed that what was
needed there was not always clear.

As they weren't well commented, most of their behavior was derived from
what made sense about a variable with that name at the time it was
coded. Because request-related code could change a lot in the future, I
though it would be better to document exactly what those variables and
events are about.

The goal is to remove as much as possible any "hidden" behavior that
seems self-evident at one time but not at a later time. One of those
hidden rule could be that there should be only one request at a given
time for a given segment.
As seen in some drafts, this may not be always true, so we should
consider any behavior relying on this as an axiom as a bug.

It does add a lot of comments, and I'm unsure if this is not even making
the player harder to read, but I still prefer it that way to the way
it was before.