api: add __priv_patchLastSegmentInSidx option
This is another alternative for being able to play Canal+ streams where
the byte-range for the last segment (deduced from parsing the sidx box
of the initialization segment) is too short.

This solution doesn't try to provide a generic solution which could
technically work for other use-cases (as those necessitate much more
work and refactoring).
In this commit, I just added an un-documented property to
transportOptions (one of the options given to a loadVideo call) called
__priv_patchLastSegmentInSidx (the name being scary on purpose).

When that boolean is set to true, the RxPlayer will consider that the
last subsegment indicated in a sidx (when in a SegmentBase-based DASH
content) goes until the end of the file.

That way, instead of doing the usual range-request with a starting and
ending byte for that last segment, the RxPlayer will only indicate the
starting byte - thus indicating that we want the remainer of the whole
file, not only a subset.

Doing that could be risky as we could be downloading much more segment
data than needed, and that's why the option is named in a scary way!

Also I chose to not document that option because of two things:

    Nobody should use that API but Canal+ for playing those broken
contents.
    We reserve the right to remove that option and replace it by a
better alternative at any time without making a new major version.
    Here we should communicate clearly that this might only be a
temporary solution.
    Note that it might be harder to communicate when the option doesn't
work anymore.