parsers/isobmff: also handle `uuid` boxes in `getBoxOffsets`
The `getBoxOffsets` allows to obtain three "offsets" in an ISOBMFF file:
  1. the starting byte of a box
  2. the first byte at which the box's content begins (the byte of the
     version if we're speaking about a "FullBox").
  3. the first byte after that box (equal to the length of the ISOBMFF
     file if we're talking about the last box).

Previously, this function did not work well with `uuid` boxes as those
box also include 16 bits for the user-defined name. This was explicitely
documented.

This is anyway never needed, but still, the implementation looked very
straightforward. I decided to add it so this exception can be removed
from the function's documentation.

I also saw a small mistake in that function where we could be returning
an offset even if the box was not found if the end of the ISOBMFF file
was truncated. This is now fixed.