compat/eme: filter out CENC with a pssh version different than 1
At Canal+, we encountered several contents where a malformed CENC pssh
box can be found. Such contents are not playable by the RxPlayer, a
 `KEY_GENERATE_REQUEST_ERROR` fatal error is triggered as soon as the
player calls the `MediaKeySession.generateRequest` API.
The only problem with that pssh box is that it is anounced with a
version of "0" despite having the format of a version "1".

Sadly, the corresponding people might be very slow to fix that problem
and even when fixed, it is ensure that previously encrypted content will
be fixed.

There is some hope of being able to cleanly fix that issue on the client
side though. The w3c note on "cenc"
(https://www.w3.org/TR/eme-initdata-cenc/), seems to say that cenc pssh
boxes are expected to always be of version "1".

We could thus theorically fix that content error by just respecting the
spec and filtering out CENC pssh boxes which are not of version 1.

However, I am afraid of breaking previously-working contents. Browsers
and CDM do not always follow those notes and recommendations, so that's
not exactly what I've done.

In this commit, I only remove CENC pssh boxes which are not of version 1,
if a CENC pssh with a version of 1 is also found in the same
initialization data.

The downside of that new logic is that it becomes very Canal+-specific:
it only fixes contents where both a valid and invalid cenc pssh boxes
are encountered.
Even if it should break no content, it adds some complexity (and a small
delay - even if it shouldn't be noticeable because segment requests are
usually done at the same time) to all contents.

Moreover, I decided to generalize the behavior of putting the CENC pssh
box(es) at the end of the concatenation of all pssh boxes given to the
`MediaKeySession.generateRequest` API to simplify the code organization.
It was only a logic performed on IE and Edge before, as those were the
only browser we know of where putting a CENC box first could lead to
problem.

This new behavior should (hopefully) not break on any other browser or
CDM.
As far as I know, there is no enforced order on pssh boxes (and browsers
are supposed to only communicate those compatible with the CDM).