tests: add integration tests for filtering CENC PSSHs with a wrong version
To better ensure that the work-around on CENC ppsh boxes with a bad
version keep working, I decided to add some integration tests.

Note that those are the first integration tests doing any kind of
EME-related calls.

To maximize the amount of browsers on which this test can work I:

  - mainly ask for a "clear key" system. This key system is required by
    the EME recommendation to be implemented on all browsers which
    supports the EME APIs.

  - to still support browsers that might implement EME APIs but might
    still not implement clear key, I also added widevine and playready
    key systems as fallbacks.

  - no licence fetching will be done. We only go until the
    `generateRequest` API, which is a stubbed call in my test.
    This means that only the following EME APIs will actually be called:
      - navigator.requestMediaKeySystemAccess
      - MediaKeySystemAccess.createMediaKeys
      - MediaKeys.createSession
      - HTMLMediaElement.setMediaKeys

  - I check that a reference to `MediaKeys` in window exists. If that's
    not the case, I signal it through a console.warn and do not run the
    test.

The `generateRequest` API is then stubbed and we check that a CENC pssh
with a version of "0" is filtered out from the initialization data, as
now expected.