utils: rewrite string parsing utils, to properly handle UTF-8 and UTF-16
For a long time, most of the string parsing utils functions in the
RxPlayer mostly assumed that handled strings contained only ASCII
characters.

We had for example the generically named: `strToBytes` and `bytesToStr`
functions which only worked for ASCII characters yet was used where
UTF-8 was the expected string encoding.
Since ASCII stays compatible with UTF-8, this worked most of the time
but it can technically break on non-ASCII characters.

A particularly badly named function was `bytesToUTF16Str`. What this
function actually did is take a little-endian UTF-16 set in an
`Uint8Array` and return the corresponding JS string.
To better reflect what it does, I decided to rename it `leUtf16ToStr`.

While doing all those changes, I also realized that an Uint16Array's
data was actually in a little-endian ordering which was not what I was
expecting.
To avoid further mistakes, I decided to be stricter on the type of
TypedArray that can be communicated through most of the RxPlayer's
internal code, by banning most explicit `Uint16Array` reference in the
code, especially as a possible function argument.

Because the demo also had similar functions and because today we cannot
share code between the RxPlayer and the demo. I decided for now to
duplicate the code in both places. I understand that it might be
considered ugly.