smooth: skip discontinuity even when we're in the middle of it
There was a strange condition in smooth content handling that made it
impossible to skip stream discontinuities when the user seeked at the
middle of one.

Basically, the player needed to still be in the segment preceding the
discontinuity to be able to skip it. This condition made no much sense,
and it isn't something that does exist in the related DASH
SegmentTimeline's code.

I'm consequently unsure if this was done on purpose, in which case we
would bring a regression here. In anyway, I find this not very probable
and if that was the case, that code should have been commented.