stream: rename the "Buffer" part of the code as "Stream"
This is a proposal to rename the `Buffer` part of the RxPlayer's code as
`Stream` - and of the corresponding `BufferOrchestrator`,
`PeriodBuffer`, `AdaptationBuffer` and `RepresentationBuffer` as
respectively `StreamOrchestrator`, `PeriodStream`, `AdaptationStream`
and `RepresentationStream`.

I thought that the name `Buffer` was poorly representative of that code's
main job - and was set mainly for legacy reasons.

The `Buffer` code does not store any segment in any buffer, it just
ensures that the right segment requests are performed at the right time
and then give those segments to the `source-buffer` code, which is
handling custom and natively-defined `SourceBuffers`.

The fact that that code is named that way can lead to a misunderstanding
 - where a reader could believe that those modules store media segments
in any way.
This is not the case, as soon as the segment fetchers (which is a whole
other module) has a new chunk available, it gives it to the "Buffer"
which immediately gives it to the right SourceBuffer.

It could seem "weird" to give it the name `Stream` as this was the
previous name of the code now called `Init`.
However, I think that `Init` is both a much better name for that other
module (it is the first core module loaded when a content is loaded and
it starts everything) and that `Stream` is a better name for the part
picking the right data from the network continuously and redirecting it
to the right SourceBuffer.

By the way, if any area of the code should be called `Buffer` it is the
code handling SourceBuffers, as we do have an abstraction buffering
media segments there waiting in a queue to be "pushable" to the
SourceBuffer (that code is in the `QueuedSourceBuffer` abstraction).
I do not think this a good idea to do that though.